+++ src/main/java/kr/ac/korea/budding/repository/FriendshipRepository.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface FriendshipRepository extends JpaRepository<FriendshipEntity, Integer> {
-    List<FriendshipEntity> findByUser1IdOrUser2Id(Integer user1Id, Integer user2Id);
+public interface FriendshipRepository extends JpaRepository<FriendshipEntity, Long> {
+    List<FriendshipEntity> findByUser1IdOrUser2Id(Long user1Id, Long user2Id);
 
-    FriendshipEntity findByUser1IdAndUser2Id(Integer user1Id, Integer user2Id);
+    FriendshipEntity findByUser1IdAndUser2Id(Long user1Id, Long user2Id);
 }
